// migrations/snapshots/c0dae822c48125bfa53e7583304b96c725727a0609bab776d2d8de75d9168dd3/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Account: {
      readonly accountId: CodecTypes['pg/uuid@1']['output'];
      readonly provider: 'GOOGLE' | 'GITHUB';
      readonly providerAccountId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly People: {
      readonly birthDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'];
      readonly peopleId: CodecTypes['pg/uuid@1']['output'];
      readonly remainingName: CodecTypes['pg/text@1']['output'];
      readonly sex: 'MALE' | 'FEMALE' | 'NON_BINARY' | 'NONE';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly PeopleContact: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly isCurrent: CodecTypes['pg/bool@1']['output'];
      readonly peopleContactId: CodecTypes['pg/text@1']['output'];
      readonly peopleId: CodecTypes['pg/uuid@1']['output'];
      readonly phoneWithDDD: CodecTypes['pg/text@1']['output'];
    };
    readonly Role: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly entity: CodecTypes['pg/text@1']['output'];
      readonly roleId: CodecTypes['pg/uuid@1']['output'];
      readonly roleType: 'ABAC' | 'RBAC';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userCan: CodecTypes['pg/text@1']['output'];
      readonly userCannot: CodecTypes['pg/text@1']['output'];
    };
    readonly Token: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly tokenId: CodecTypes['pg/uuid@1']['output'];
      readonly type: 'PASSWORD_RECOVER';
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdOrigin: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
    readonly UserRole: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdBy: CodecTypes['pg/uuid@1']['output'];
      readonly roleId: CodecTypes['pg/uuid@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly userRoleId: CodecTypes['pg/uuid@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Account: {
      readonly accountId: CodecTypes['pg/uuid@1']['input'];
      readonly provider: 'GOOGLE' | 'GITHUB';
      readonly providerAccountId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly People: {
      readonly birthDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'];
      readonly peopleId: CodecTypes['pg/uuid@1']['input'];
      readonly remainingName: CodecTypes['pg/text@1']['input'];
      readonly sex: 'MALE' | 'FEMALE' | 'NON_BINARY' | 'NONE';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly PeopleContact: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly isCurrent: CodecTypes['pg/bool@1']['input'];
      readonly peopleContactId: CodecTypes['pg/text@1']['input'];
      readonly peopleId: CodecTypes['pg/uuid@1']['input'];
      readonly phoneWithDDD: CodecTypes['pg/text@1']['input'];
    };
    readonly Role: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly entity: CodecTypes['pg/text@1']['input'];
      readonly roleId: CodecTypes['pg/uuid@1']['input'];
      readonly roleType: 'ABAC' | 'RBAC';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userCan: CodecTypes['pg/text@1']['input'];
      readonly userCannot: CodecTypes['pg/text@1']['input'];
    };
    readonly Token: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly tokenId: CodecTypes['pg/uuid@1']['input'];
      readonly type: 'PASSWORD_RECOVER';
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdOrigin: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
    readonly UserRole: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdBy: CodecTypes['pg/uuid@1']['input'];
      readonly roleId: CodecTypes['pg/uuid@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly userRoleId: CodecTypes['pg/uuid@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly accounts: {
      readonly account_id: CodecTypes['pg/uuid@1']['output'];
      readonly provider: 'GOOGLE' | 'GITHUB';
      readonly provider_account_id: CodecTypes['pg/text@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly people_contacts: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly is_current: CodecTypes['pg/bool@1']['output'];
      readonly people_contact_id: CodecTypes['pg/text@1']['output'];
      readonly peopleId: CodecTypes['pg/uuid@1']['output'];
      readonly phone_with_ddd: CodecTypes['pg/text@1']['output'];
    };
    readonly peoples: {
      readonly birthDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'];
      readonly people_id: CodecTypes['pg/uuid@1']['output'];
      readonly remainingName: CodecTypes['pg/text@1']['output'];
      readonly sex: 'MALE' | 'FEMALE' | 'NON_BINARY' | 'NONE';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly roles: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly entity: CodecTypes['pg/text@1']['output'];
      readonly role_id: CodecTypes['pg/uuid@1']['output'];
      readonly role_type: 'ABAC' | 'RBAC';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userCan: CodecTypes['pg/text@1']['output'];
      readonly userCannot: CodecTypes['pg/text@1']['output'];
    };
    readonly tokens: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly token_id: CodecTypes['pg/uuid@1']['output'];
      readonly type: 'PASSWORD_RECOVER';
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly user_roles: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly created_by: CodecTypes['pg/uuid@1']['output'];
      readonly role_id: CodecTypes['pg/uuid@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
      readonly user_role_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly users: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly created_origin: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly password_hash: CodecTypes['pg/text@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;